[fe#37] rename: 스타일 컴포넌트 폴더구조 세분화
