+++ fe/src/views/header/gnbLists/gnbLists.styled.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const GnbMenu = styled.li`
+  margin-right: 24px;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 26px;
+  cursor: pointer;
+`;
+
+export const GnbSubMenuWrapper = styled.ul`
+  margin-top: 4px;
+`;
+
+export const GnbSubMenu = styled.li`
+  font-size: 14px;
+  line-height: 24px;
+`;
