+++ fe/src/views/header/svg/svg.js
@@ -1,4 +1,4 @@
-import { StyledSvgWrapper } from '../Header.styled';
+import { StyledSvgWrapper } from './svg.styled.js';
 
 export function Svg({ icons, useIconsName }) {
   return useIconsName.map((iconName, idx) => (
