+++ fe/src/views/header/svg/svg.styled.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const StyledSvgWrapper = styled.div`
+  margin-left: 24px;
+  width: 24px;
+  cursor: pointer;
+`;
