+++ fe/src/views/main/bestProduct/BestProduct.styled.js
@@ -31,40 +31,7 @@ export const BestProductHeader = styled.div`
   border-bottom: 1px solid #ebebeb;
 `;
 
-export const Title = styled.h1`
-  margin-left: 16px;
-  font-style: normal;
-  font-weight: 700;
-  font-size: 40px;
-  line-height: 58px;
-`;
-
 export const TabBar = styled.ul`
   ${({ flex }) => applyFlex({ flex })}
   margin-top: 24px;
-`;
-
-export const TabMenuLi = styled.li`
-  margin-right: 32px;
-  padding-bottom: 17px;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 20px;
-  line-height: 30px;
-  color: #bdbdbd;
-  border-bottom: 1px solid white;
-  cursor: pointer;
-  ${props => {
-    if (props.curTab === props['data-menu-id']) {
-      return `
-        color: #000;
-      `;
-    }
-  }}
-`;
-
-export const StyledDiv = styled.div`
-  ${({ flex }) => applyFlex({ flex })}
-  margin-top: 34px;
-  padding: 0 80px;
 `;
