+++ fe/src/views/main/bestProduct/tabList/tabList.js
@@ -1,4 +1,4 @@
-import { StyledDiv } from '../BestProduct.styled';
+import { StyledDiv } from './tabList.styled';
 import { ProductCard } from '../../../../components/ProductCard';
 
 export function TabList({ tabList }) {
