+++ fe/src/views/main/bestProduct/title/exhibitionTitle.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Title = styled.h1`
+  margin-left: 16px;
+  font-style: normal;
+  font-weight: 700;
+  font-size: 40px;
+  line-height: 58px;
+`;
